ai-pipeline/requirements.txt
openai
requests
PyYAML
pytest
GitPython
Flask
anthropic
